utils/requirements.txt
psycopg2-binary
python-logging-loki
